fix: Store using the correct schema when migrating from stable memory to map (#4626)
# Motivation
There is a bug:
- In the pre-upgrade hook, we check the current schema and save state
accordingly. That is fine.
- But we determine the schema by looking at the stable memory, not by
checking which schema is currently in use.
- When migrating from the heap to stable structures, the stable
structures are written to stable memory, so this gives the correct
result IF migration has completed.
- But when migrating from stable structures to map, the migration
completes correctly, but then in the `pre_upgrade` hook we look at the
stable memory, which is still set up for stable structures, and say "oh,
save to stable structures". This not only un-does the migration but also
loses any data added after the migration was completed.
- It also gives the incorrect result part way through a migration from
map to stable structures.

# Changes
- Use the correct schema when saving data in the pre-upgrade hook.

# Tests
- The bug can be exercised by running the existing migration test twice,
with more accounts in the second run.
- If we test that on main, [the test
fails](https://github.com/dfinity/nns-dapp/actions/runs/8324594115).
  - But in this PR, the same test succeeds.

# Todos

- [ ] Add entry to changelog (if necessary).