NNS1-2918: Add settings button to tokens table (#4631)
# Motivation

We want to allow hiding tokens with a zero balance from the tokens
table.
For this there will be a settings button which will open a popup with a
toggle.
This PR only adds the settings button when the feature flag is enabled.
The button doesn't do anything yet.

# Changes

1. Add a `header-icon` slot in `TokensTable` to place the settings
button in.
2. Use `height` instead of vertical `padding` so the table header looks
the same with and without the settings button.
3. Use the `header-icon` slot in `frontend/src/lib/pages/Tokens.svelte`
to place the settings button.
4. Make sure the button is the correct style when active.

# Tests

Added unit test that the button is only there when the feature flag is
enabled.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet