Mock queryFinalizationStatus in ProdLaunchpad.spec.ts (#4643)
# Motivation

We have a copy of (the 3 pages of) the prod sns aggregator data here:
https://github.com/dfinity/nns-dapp/tree/main/frontend/src/tests/workflows/Launchpad
We use this in `ProdLaunchpad.spec.ts` to test that we can correctly
parse the prod SNS data.

We also have a weekly job to keep this data up-to-date.
The [most recent update
PR](https://github.com/dfinity/nns-dapp/pull/4641) caused
`ProdLaunchpad.spec.ts` [to
fail](https://github.com/dfinity/nns-dapp/actions/runs/8369291084/job/22914793727?pr=4641).

This is because if an SNS was closed less than a week ago, we try to
fetch its finalization status and this call was not mocked. And this is
the first time since the automatic update job was added that SNSes
(ICLighthouse DAO and ELNA AI) were completed.

# Changes

Mock `queryFinalizationStatus` in `ProdLaunchpad.spec.ts`.

# Tests

Tested on branch `bot-aggregator-response-update-1710991986` of
https://github.com/dfinity/nns-dapp/pull/4641 that this makes the test
pass again.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary