Make rowHref on token data non-optional (#4647)
# Motivation

Token table rows are `<a>` tags most of the time but in some cases (when
token data or account details are still loading) we don't provide an
`href` and make the element a `<div>`.
But there is no reason we can't provide an `href` in those cases as
well.
And if we do the code is simpler and the rows are always clickable.

An additional benefit is that if every row has a unique `href` we could
use this as key in the `{#each}` which would be necessary to provide
smooth transitions when hiding zero balance tokens.

# Changes

1. Make `rowHref` mandatory on `UserTokenLoading` and `UserTokenData`.
2. Make sure `rowHref` is always provided.
3. Make the row always an `<a>` tag.

# Tests

First I provided `rowHref` in all the tests as required by the
compliler.
Then I replaced the newly provided `rowHref` value with empty strings
and verified that the tests now fail because of the newly expected
values.

For manual testing I added a delay in loading token data and account
details and verified that while loading the row is now clickable while
previously it wasn't.

# Todos

- [x] Add entry to changelog (if necessary).