Consistently exclude account param from URL for main accounts (#4651)
# Motivation

We allow navigating to token wallets when not signed in.
When the user is not signed in, we don't know the account identifier so
we don't include it in the URL.
When the user signs in, if there is no account identifier in the URL, we
show the main account for that token by default.
This means that the URL you get when navigating to a main account is
different depending on whether you were signed in when you navigated.
This is not necessary.

This PR makes the experience more consistent and URLs cleaner by always
omitting the `account` param when navigating to a main account.

Additionally, when using the `rowHref` as the Svelte `{#each}` key, it
means that the row for the ICP main account is considered the same
before and after signing in, while previously it would have been
considered a new row.

# Changes

When creating a wallet URL, omit the `account` param if it would have
been for the main account.

# Tests

1. Updated unit tests to not expect account parameter for main account.
2. Tested manually to see that the account identifier is only present in
URL for sub/hardware wallet accounts.
3. Also tested manually in another branch to see that the animation
doesn't replace one Main row with another as it did before.

# Todos

- [x] Add entry to changelog (if necessary).