test: Check that stats are not recomputed (#4616)
# Motivation
Until the least release, account stats such as the total number of
hardware wallets was computed as part of every upgrade in the
`post_upgrade` hook. Iterating over all accounts on the heap was
acceptably fast. This is absolutely not the case when accounts are
stored in stable memory. Using the same code path when stable memory is
in use would cause the upgrade to fail. We therefore wish to:

- Test that stats are not recomputed on upgrade any more, during normal
clean upgrades.
- Delete the code for recomputing stats once we are sure that the new
state of affairs is stable.

# Changes
- Make it visible whether stats were recomputed during an upgrade by
recording this in the public stats.
- Add a test that, with normal downgrades and upgrades, stats are never
recomputed.

# Tests
- A test is included.

# Todos

- [x] Add entry to changelog (if necessary).