Increase the number of acounts used in randomized tests (#4661)
# Motivation
As part of the security review with @roelstorms, it was considered that
we probably want to increase the number of accounts used in randomized
migration testing.

The number of accounts is a random number in the range 0..40. If we
migrate 20 accounts per step (increased from 10), half the time the
entire migration is completed in one step. (Plus a second step to point
the switch at the new database.) This gives insufficient opportunity for
the test to find any bugs.

# Changes
- Increase the number of accounts in randomized migration tests to a
random number in the range 0..100.

# Tests
- See CI

# Todos

- [ ] Add entry to changelog (if necessary).