Fix copied tooltip IDs (#4690)
# Motivation

A tooltip appears when you hover over a tooltip target.
The tooltip target should have an `aria-describedby` attribute that
matches the `id` attribute on the tooltip such that screen readers can
associate the two.
And for this for work correctly, the `id` attribute has to be unique.

But I found some tooltip ID which were copied and not unique.

# Changes

1. Fix the incorrectly copied tooltip IDs to match their context.
2. Add unit tests for the affected tooltip IDs.
3. Add getters on page objects where necessary.

# Tests

Unit tests added.

Also, this was caught by another change where the page object getter
actually depends on this ID and that caused some tests to fail and this
fixed them.

# Todos

- [x] Add entry to changelog (if necessary).