Get tooltip element in page object via id (#4699)
# Motivation

There are changes in the "next" version of gix-components that place
tooltip elements directly in the body to avoid issues with
`overflow:hidden`.
With the current implementation of `TooltipPo` this will prevent the
page object from finding the tooltip element where it used to be. The
solution is to use the `aria-describedby` attribute on the tooltip
target to get the `id` attribute of the tooltip element and query for
the element using that `id`.
This should work even without the new tooltip implementation.

# Changes

1. In `TooltipPo`, use the `aria-describedby` attribute to find the
tooltip element by ID in the document body.
2. Add a `getDocumentBody` method to `PageObjectElement` to have an
abstracted way to find an element anywhere in the body. Implement it in
`JestPageObjectElement` but mark it as not implemented in
`PlaywrightPageObject` (we'll implement it when we need it).
3. In `frontend/src/tests/lib/components/ui/IdentifierHash.spec.ts` make
sure the container used for the page object is actually in the document
body so the page object will be able to find the tooltip element.

# Tests

pass

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary