Split VotingNeuronListItem from VotingNeuronSelectList (#4710)
# Motivation

The `VotingNeuronSelectList` has a list of items, one for each neuron.

<img width="789" alt="Screenshot 2024-04-10 at 11 39 00"
src="https://github.com/dfinity/nns-dapp/assets/122978264/531532aa-4113-4d58-a1a1-1afa2c00f8dc">

To make it easier to test these items in isolation, we make a separate
component for the item.

# Changes

Create a new `VotingNeuronListItem` component and use it in
`VotingNeuronSelectList` instead of rendering the items directly in the
list.

# Tests

Unit test added for the new `VotingNeuronListItem` component.
Unit test for `VotingNeuronSelectList` continue to pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary