Use page objects in VotingNeuronSelectList (#4711)
# Motivation

Making the test easier to read and maintain.

Also 2 tests are nonsense and are deleted. These tests assume that the
component shows a sum of voting powers but `VotingNeuronSelectList`
doesn't do that at all. The sum of selected voting powers is rendered in
`VotableNeuronList` instead.

`"should not display total voting power of neurons"` tests that the
total voting power of neurons (as opposed to the total voting power of
ballots) is not displayed. But neither is displayed so obviously this
passes but it's meaningless.
`"should recalculate total voting power after selection"` doesn't test
anything because the only expectation in the test is an async
expectation that's not awaited. If `await` is added to the test, the
test fails because the component doesn't render any sum of voting
powers.

# Changes

1. Add `VotingNeuronSelectListPo` to use in the test.
2. Add `allUnder` to `VotingNeuronListItemPo` to get a list of page
objects, used in `VotingNeuronSelectListPo`.
3. Use `VotingNeuronSelectListPo` in `VotingNeuronSelectList.svelte`.

# Tests

only tests

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary