NNS1-2906: Remove the code path for when ENABLE_ICP_INDEX is disabled (#4714)
# Motivation

`ENABLE_ICP_INDEX` is enabled on mainnet and in tests so the negative
code path is not used anymore.

# Changes

Remove the conditionals on `ENABLE_ICP_INDEX` and only keep the branch
where `ENABLE_ICP_INDEX` is true.

All the code that becomes unused as a result will be cleaned up in
future PRs.

# Tests

Still pass

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: georgi-dfinity <112886364+georgi-dfinity@users.noreply.github.com>