NNS1-3003: Rename topics and proposal types (#4726)
# Motivation

We are making changes to topics and proposals described in [this forum
post](https://forum.dfinity.org/t/bringing-clarity-to-icp-upgrade-proposals/29626).
Some topics/proposals are deprecated and new ones made in their place.
Others are reused and renamed.
This PR changes the names of the reused ones.

# Changes

1. Change topic "Replica Version Management" to "IC OS Version Election"
2. Change topic "Subnet Replica Version Management" to "IC OS Version
Deployment"
3. Change proposal type "Update Subnet Replica Version" to "Deploy
GuestOS For All Subnet Nodes"
4. Change proposal type "Update Elected Replica Versions" to "Revise
Elected GuestOS Version".

# Tests

Updated

# Todos

- [x] Add entry to changelog (if necessary).