fix: Replace the downgrade-upgrade test with a call to the migration test (#4637)
# Motivation
The downgrade-upgrade test verifies that the nns-dapp canister, as
proposed in a PR, can be downgraded to the production build and that we
can upgrade from the production build to the current build. The problem:
The current test uses the same canister arguments to deploy the
production Wasm as it does to deploy the PR Wasm. This makes the
downgrade-upgrade test ineffective if the main difference between one
release and the proposed next release is a change in the arguments. For
example, schema migration can be done entirely by deploying with
different arguments.

Options:
- We _could_ adapt the existing downgrade-upgrade test to accept a
second argument file and use it.
- On the other hand, the migration test is already a copy of the
downgrade-upgrade test, adapted to allow it to test any pair of Wasms
and arguments, not just prod and the current code.

# Changes
- Use the migration test in place of the downgrade-upgrade test.
- The migration test is run against the prod Wasm, as before, but
additionally uses the arguments for local deployment published with the
production Wasm.
  - Use `dfx start` instead of a snapshot, because:
    - This matches other migration tests
- While the snapshot should work, it takes more time setting up
canisters we don't need.
- Fix an issue in the migration test where a test for an undefined
schema crashes on an undefined schema.

# Tests
- See CI

# Todos

- [ ] Add entry to changelog (if necessary).