Fix confusing test output caused by missing mocks (#4725)
# Motivation

Some of our tests produce output like this:
```
TypeError: id.transformRequest is not a function
    at HttpAgent.call (/Users/dskloet/dev/nns-dapp/tree9/frontend/node_modules/@dfinity/agent/src/agent/http/index.ts:424:33)
    at caller (/Users/dskloet/dev/nns-dapp/tree9/frontend/node_modules/@dfinity/agent/src/actor.ts:484:28)
    at getTransactions (/Users/dskloet/dev/nns-dapp/tree9/frontend/node_modules/@dfinity/ledger-icp/dist/esm/chunk-SUVDEL6N.js:3:3602)
    at Module.getTransactions (/Users/dskloet/dev/nns-dapp/tree9/frontend/src/lib/api/icp-index.api.ts:35:37)
    at Module.loadIcpAccountTransactions (/Users/dskloet/dev/nns-dapp/tree9/frontend/src/lib/services/icp-transactions.services.ts:23:42)
    at reloadTransactions (/Users/dskloet/dev/nns-dapp/tree9/frontend/src/lib/pages/NnsWallet.svelte:1245:3)
    at accountDidUpdate (/Users/dskloet/dev/nns-dapp/tree9/frontend/src/lib/pages/NnsWallet.svelte:1272:3)
```
Because a function (in this case `getTransactions`) is called that
should have been mocked out but wasn't.
This doesn't cause the test to fail because the test tests something
that doesn't depend on this call, and the test passes even before these
errors are logged.
If we introduce an artificial delay, these tests fail because tests that
log anything are made to fail.
So I temporarily added a delay after every test to find all these tests
and fix them.

# Changes

Add missing mock behavior to get rid of the confusing test logs.

# Tests

Pass, even with the artificial delay.
But I've removed the artificial delay again because we don't want tests
to be slow.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary