NNS1-2905: Delete unused get_transactions and related code (#4749)
# Motivation

NNS dapp now reads transactions from the index canister and no longer
from the nns-dapp canister.
In https://github.com/dfinity/nns-dapp/pull/4723 we removed the canister
method for retrieving transactions.
This caused some dead code that was not yet removed.
This PR removes `get_transactions` on the `AccountStore` are related
dead code as well as tests for it.
The transactions themselves still exist in the account store and will
have to be removed with a separate migration.

# Changes

1. Removed method `AccountStore::get_transactions`.
2. Remove types `GetTransactionsRequest`, `GetTransactionsResponse`,
`TransactionResult`, `TransferResult`
3. Remove tests that depend on `get_transactions`.
4. Remove no longer used `get_queue_for_testing` and
`get_mut_queue_for_testing` from `MultiPartTransactionToBeProcessed`.

# Tests

Removed.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary