Remove explicit chrono dependency (#4750)
# Motivation

`rs/backend/Cargo.toml` has an explicit dependency on `chrono` version
0.4.19.
But no code in nns-dapp depends directly on `chrono` (nor did it when
the dependency was added).
Removing the explicit dependency does not seem to break anything.
If I increase the explicit dependency to 0.4.20 (or higher), I do get
the error from the comment but not if I just remove it.
And having the dependency prevents us from updating other dependencies
which require different versions of `chrono`.

# Changes

1. Remove the `chrono` dependency from `rs/backend/Cargo.toml`.

# Tests

CI passes

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary