NNS1-2905: Stop appending new transactions in storage (#4759)
# Motivation

We no longer read transactions from the nns-dapp canister, but from the
index canister instead.
We want to remove the transactions from nns-dapp canister so we need to
remove all the code that is still using the transactions.

However, we do still need to process the transactions to check if we
need to continue claiming or topping up a neuron or canister.

In this PR we stop appending transactions to storage and instead only
process them.

# Changes

1. Rename`append_transaction` to `maybe_process_transaction`. It used to
both append and process the transaction but now it only processed the
transaction.
2. Remove the logic to append the transaction.
3. Pass `transfer` by reference because lint complained that it's not
consumed.
4. Rename `try_add_transaction_to_account` to `store_has_account` and
remove most of the functionality. `try_add_transaction_to_account` used
to return whether the transaction was added, which was the case if we
had an account with that identifier in the store. We still need to check
if we have the account but no longer add the transaction.

# Tests

Update the tests to stop checking that the transaction was added and
only check that it is processed.

The test for transaction pruning has been removed because there is
nothing to prune if nothing is added. The pruning code itself will be
removed shortly.

# Todos

- [x] Add entry to changelog (if necessary).